backend/requirements.txt
livekit-agents==1.3.11
livekit-plugins-openai==1.3.11
livekit-plugins-deepgram==1.3.11
livekit-plugins-cartesia==1.3.11
livekit-plugins-silero==1.3.11
livekit==1.0.23
supabase
python-dotenv
aiohttp
